add grass.py and the class into
